pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "datasets>=4.8.4",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pyyaml>=6.0.3",
    "tensorboard>=2.20.0",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
]
